+++ lib/output/redis_list.go
@@ -17,9 +17,16 @@ func init() {
 		Summary: `
 Pushes messages onto the end of a Redis list (which is created if it doesn't
 already exist) using the RPUSH command.`,
+		Description: `
+The field ` + "`key`" + ` supports
+[interpolation functions](/docs/configuration/interpolation#bloblang-queries), allowing
+you to create a unique key for each message.`,
 		Async: true,
 		FieldSpecs: redis.ConfigDocs().Add(
-			docs.FieldCommon("key", "The key of a Redis list."),
+			docs.FieldCommon(
+				"key", "The key for each message, function interpolations can be optionally used to create a unique key per message.",
+				"benthos_list", "${!meta(\"kafka_key\")}", "${!json(\"doc.id\")}", "${!count(\"msgs\")}",
+			).SupportsInterpolation(false),
 			docs.FieldCommon("max_in_flight", "The maximum number of messages to have in flight at a given time. Increase this to improve throughput."),
 		),
 		Categories: []Category{